+++ rook/etc/roocs.ini
@@ -1,0 +1,17 @@
+[project:cmip5]
+inventory_url =
+
+[project:cmip6]
+inventory_url =
+
+[project:cordex]
+inventory_url =
+
+[project:c3s-cmip5]
+inventory_url =
+
+[project:c3s-cmip6]
+inventory_url = https://raw.githubusercontent.com/cp4cds/c3s_34g_manifests/add_inventories/inventories/c3s-cmip6/c3s-cmip6_files_latest.yml
+
+[project:c3s-cordex]
+inventory_url =
