+++ rook/utils/average_utils.py
@@ -3,12 +3,9 @@ from copy import deepcopy
 from .input_utils import resolve_collection_if_files
 
 
-def run_average(args):
+def run_average(kwargs):
     # Convert file list to directory if required
-    kwargs = deepcopy(args)
-    original_collection = args.get("collection")
-    kwargs["collection"] = resolve_collection_if_files(original_collection)
-
+    original_collection = kwargs.get("collection")
     # TODO: handle lazy load of daops
     # from daops.ops.average import average
